requirements.txt
Django==2.1.3
django-pyodbc-azure==2.1.0.0
pyodbc==4.0.24
pytz==2018.7
djangorestframework==3.9.0
django-filter==2.0.0
Django==2.1.3
django-configurations==2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
Pillow==5.3.0